requirements.txt
django
djangorestframework
djangorestframework-simplejwt
mysqlclient
drf-spectacular
django-environ